+++ NEWS.md
@@ -1,7 +1,7 @@
 # DEbPeak 1.0.1
 ## New features
 * Added `DiffPeakPie` to stat genomic regions of differential peaks with pie plot.
-
+* Added `EnrichPlot` to create bar or dot plot for functional enrichment analysis, this is useful for visualizing selected terms.
 
 -------------